fix compile error

+++ include/platform/windows/program.hpp
@@ -128,7 +128,7 @@ inline uint32_t get_self_pid() {
     return GetCurrentProcessId();
 }
 
-inline void set_cgroup_cpu_limit(std::error_code& ec, int) {
+inline void set_cgroup_cpu_limit(std::error_code& ec, float) {
     ec.clear();
 }
